pyproject.toml
[project]
name = "testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "email-validator>=2.3.0",
    "fastapi>=0.118.0",
    "greenlet>=3.2.4",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pydantic>=2.11.10",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-json-logger>=4.0.0",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.43",
    "structlog>=25.4.0",
    "uvicorn>=0.37.0",
]
